microservices/reports/requirements.txt
accelerate==1.6.0
aiofiles==24.1.0
alembic==1.18.4
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.11.1
attrs==25.3.0
bidict==0.23.1
blinker==1.9.0
brotli==1.2.0
cachetools==5.5.2
certifi==2025.4.26
cffi==2.0.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
ConfigArgParse==1.7.1
contourpy==1.3.1
cycler==0.12.1
Django==5.2.11
djangorestframework==3.16.1
dnspython==2.8.0
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
Flask==3.1.2
flask-cors==6.0.2
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-PyMongo==3.0.1
Flask-SQLAlchemy==3.1.1
fonttools==4.56.0
fsspec==2025.3.2
future==1.0.0
gevent==25.9.1
geventhttpclient==2.3.9
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc0
google-api-python-client==2.168.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
gradio==5.27.0
gradio_client==1.9.0
greenlet==3.3.1
groovy==0.1.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
locust==2.43.3
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
MiniSom==2.3.5
mpmath==1.3.0
msgpack==1.1.2
mysqlclient==2.2.8
narwhals==1.44.0
networkx==3.4.2
numpy==2.2.3
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==6.2.0
pluggy==1.5.0
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
psycopg2-binary==2.9.11
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.11.3
pydantic_core==2.33.1
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
pymongo==4.16.0
PyMySQL==1.1.2
pyparsing==3.2.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.13.1
python-multipart==0.0.20
python-socketio==5.16.1
pytz==2025.1
pywin32==311
PyYAML==6.0.2
pyzmq==27.1.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
ruff==0.11.7
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.2
seaborn==0.13.2
semantic-version==2.10.0
serializers==0.2.4
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.46
sqlparse==0.5.5
starlette==0.46.2
streamlit==1.46.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.1
toml==0.10.2
tomlkit==0.13.2
torch==2.7.0
tornado==6.5.1
tqdm==4.67.1
transformers==4.51.3
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
watchdog==6.0.0
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.5
wsproto==1.3.2
zope.event==6.1
zope.interface==8.2
